// src/db/node-types/videos/types/InputVideoCreate.ts
export type InputVideoCreate = {
    video_provider: string
    external_id: string
    title: string | null
    description: string | null
    creator: string | null
    license: string | null
    tags: string | null
    source: string | null
    duration: string | null
    thumbnail_url_l: string | null
    thumbnail_url_m: string | null
    thumbnail_url_s: string | null
    thumbnail_url_xs: string | null
}